+++ src/org/thoughtcrime/securesms/ConversationListFragment.java
@@ -238,6 +238,7 @@ public class ConversationListFragment extends Fragment
 
       @Override
       protected void onPostExecute(Optional<? extends Reminder> reminder) {
+        DozeReminder.maybeAskDirectly(getActivity());
         if (reminder.isPresent() && getActivity() != null && !isRemoving()) {
           reminderView.showReminder(reminder.get());
         } else if (!reminder.isPresent()) {