+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Delta Chat Android Changelog
 
+## v1.12.5
+2020-08
+
+* fix notifications for Android 4
+* fix and streamline querying permissions
+* fix removing POIs from map
+* fix emojis displayed on map
+* fix: connect directly after account qr-scan
+* make bot-commands such as /echo clickable
+
+
 ## v1.12.3
 2020-08